# core/mount: use ptrace instead of go:linkname