@fuweid 2024-08-19
>     default: time="2024-08-19T14:13:35.673075415Z" level=fatal msg="serve failure" address=/run/containerd-test/containerd.sock error="accept unix /run/containerd-test/containerd.sock: accept4: bad file descriptor"

Hmm. Never see this error before...