@cpuguy83 2024-08-26
Seems like this is making it so this will only work on kernel >= 5.4.
Seems like we should have a less optimized version that doesn't require pidfd?
Maybe a full re-exec and pass fd back over unix socket?