@fuweid 2024-08-27
> Seems like this is making it so this will only work on kernel >= 5.4. Seems like we should have a less optimized version that doesn't require pidfd? Maybe a full re-exec and pass fd back over unix socket?

The re-exec solution is complicated because it requires all the binaries implement the command to support re-exec.
Both containerd and containerd-shim have to do that. 

For this case, we don't need to run a real command. With `ptrace`, the `execve(2)` syscall will be stopped at exit-to-user loop. And then, if kubernetes users want to use user-namespace feature, at least, they should have >= v6.3 linux kernel.
The IDMapped mount is the key to have better experience for user-namespace. That's why I think pidfd is better than re-exec. 

From https://kubernetes.io/docs/concepts/workloads/pods/user-namespaces/ :

```
This is a Linux-only feature and support is needed in Linux for idmap mounts on the filesystems used. This means:

On the node, the filesystem you use for /var/lib/kubelet/pods/, or the custom directory you configure for this, needs idmap mount support.

* All the filesystems used in the pod's volumes must support idmap mounts.

* In practice this means you need at least Linux 6.3, as tmpfs started supporting idmap mounts in that version. This is usually needed as several Kubernetes features use tmpfs (the service account token that is mounted by default uses a tmpfs, Secrets use a tmpfs, etc.)

* Some popular filesystems that support idmap mounts in Linux 6.3 are: btrfs, ext4, xfs, fat, tmpfs, overlayfs.

* In addition, the container runtime and its underlying OCI runtime must support user namespaces. The following OCI runtimes offer support:

[crun](https://github.com/containers/crun) version 1.9 or greater (it's recommend version 1.13+).
```