@kolyshkin 2024-11-14
Related to this: https://github.com/golang/go/issues/70352